Correct syntax of .readme
I changed  "and leaving running" to "and leave running" to make the grammar consistent throughout the doc. I'm excited to create my own spee.ch server for hosting citizen science images for training collective intelligence AI. I love the idea of the library returning as an epicenter for all knowledge and art to mingle and in so-doing, become permanent without risk of censorship or destruction.